fix: accept UINT32 input instead of FLOAT32 in Vulkan backend
Correct the input type whitelist to match the conversion shader.
Add regression coverage for UINT32 output and FLOAT32 rejection.

+++ src/backends/vulkan/vulkan_backend.cpp
@@ -1836,7 +1836,7 @@ void VulkanBackend::initialize(const ProcessorConfiguration& config) {
 		case IOBuffer::DataType::UINT16:
 			this->impl->bytesPerSample = 2;
 			break;
-		case IOBuffer::DataType::FLOAT32:
+		case IOBuffer::DataType::UINT32:
 			this->impl->bytesPerSample = 4;
 			break;
 		default:
